Use start timestamp from vosk youtube

+++ src/helper/vosk.ts
@@ -1,8 +1,8 @@
 export const typedVoskResponse = (data: any): VOSKResponse => {
   const typed = JSON.parse(data)
   typed.listen = typed.listen === 'true'
-  typed.start = typed.start ? parseFloat(typed.start) : undefined
-  typed.stop = typed.stop ? parseFloat(typed.stop) : undefined
+  typed.start = typed.start ? parseInt(typed.start) : undefined
+  typed.stop = typed.stop ? parseInt(typed.stop) : undefined
 
   return typed
 }
